+++ packages/message-sdk/src/components/UserList.tsx
@@ -1,29 +1,20 @@
-import {
-  sendFriendRequest,
-  unFriend,
-} from "@coral-xyz/app-extension/src/api/friendship";
-import type {
-  RemoteUserData} from "@coral-xyz/common";
-import {
-  NAV_COMPONENT_MESSAGE_PROFILE
-} from "@coral-xyz/common";
+import type { RemoteUserData } from "@coral-xyz/common";
+import { NAV_COMPONENT_MESSAGE_PROFILE , sendFriendRequest, unFriend } from "@coral-xyz/common";
 import { updateFriendshipIfExists } from "@coral-xyz/db";
-import {
-  isFirstLastListItemStyle,
-  PrimaryButton,
-  ProxyImage,
-} from "@coral-xyz/react-common";
-import { useUser } from "@coral-xyz/recoil";
-// import { useNavigation } from "@coral-xyz/recoil";
+import { isFirstLastListItemStyle, ProxyImage } from "@coral-xyz/react-common";
+import { useNavigation, useUser } from "@coral-xyz/recoil";
 import { useCustomTheme } from "@coral-xyz/themes";
-import ArrowForwardIcon from "@mui/icons-material/ArrowForward";
 import { List, ListItem } from "@mui/material";
-
-import { ParentCommunicationManager } from "../ParentCommunicationManager";
 
 import { useStyles } from "./styles";
 
-export const UserList = ({ users }: { users: RemoteUserData[] }) => {
+export const UserList = ({
+  users,
+  setMembers,
+}: {
+  users: RemoteUserData[];
+  setMembers?: React.Dispatch<React.SetStateAction<RemoteUserData[]>>;
+}) => {
   const theme = useCustomTheme();
 
   return (
@@ -37,10 +28,10 @@ export const UserList = ({ users }: { users: RemoteUserData[] }) => {
     >
       {users.map((user, index) => (
         <UserListItem
-          key={user.id}
           user={user}
           isFirst={index === 0}
           isLast={index === users.length - 1}
+          setMembers={setMembers}
         />
       ))}
     </List>
@@ -51,12 +42,15 @@ function UserListItem({
   user,
   isFirst,
   isLast,
+  setMembers,
 }: {
   user: RemoteUserData;
   isFirst: boolean;
   isLast: boolean;
+  setMembers?: React.Dispatch<React.SetStateAction<RemoteUserData[]>>;
 }) {
   const theme = useCustomTheme();
+  const { push } = useNavigation();
   const classes = useStyles();
   const { uuid } = useUser();
   return (
@@ -64,7 +58,7 @@ function UserListItem({
       button
       disableRipple
       onClick={() => {
-        ParentCommunicationManager.getInstance().push({
+        push({
           title: `@${user.username}`,
           componentId: NAV_COMPONENT_MESSAGE_PROFILE,
           componentProps: {
@@ -94,7 +88,6 @@ function UserListItem({
         }}
       >
         <div
-          className={classes.hoverParent}
           style={{ flex: 1, display: "flex", justifyContent: "space-between" }}
         >
           <div style={{ display: "flex" }}>
@@ -111,29 +104,75 @@ function UserListItem({
           </div>
           <div>
             <div
-              onClick={async () => {
+              onClick={async (e) => {
+                e.stopPropagation();
                 if (user.areFriends) {
                   await unFriend({ to: user.id });
                   await updateFriendshipIfExists(uuid, user.id, {
                     areFriends: 0,
                     requested: 0,
                   });
+                  setMembers?.((members) =>
+                    members.map((m) => {
+                      if (m.id === user.id) {
+                        return {
+                          ...m,
+                          areFriends: false,
+                        };
+                      }
+                      return m;
+                    })
+                  );
                 } else if (user.requested) {
                   await sendFriendRequest({ to: user.id, sendRequest: false });
                   await updateFriendshipIfExists(uuid, user.id, {
                     requested: 0,
                   });
+                  setMembers?.((members) =>
+                    members.map((m) => {
+                      if (m.id === user.id) {
+                        return {
+                          ...m,
+                          requested: false,
+                        };
+                      }
+                      return m;
+                    })
+                  );
                 } else if (user.remoteRequested) {
                   await sendFriendRequest({ to: user.id, sendRequest: true });
                   await updateFriendshipIfExists(uuid, user.id, {
                     requested: 0,
                     areFriends: 1,
                   });
+                  setMembers?.((members) =>
+                    members.map((m) => {
+                      if (m.id === user.id) {
+                        return {
+                          ...m,
+                          requested: false,
+                          areFriends: true,
+                        };
+                      }
+                      return m;
+                    })
+                  );
                 } else {
                   await sendFriendRequest({ to: user.id, sendRequest: true });
                   await updateFriendshipIfExists(uuid, user.id, {
                     requested: 1,
                   });
+                  setMembers?.((members) =>
+                    members.map((m) => {
+                      if (m.id === user.id) {
+                        return {
+                          ...m,
+                          requested: true,
+                        };
+                      }
+                      return m;
+                    })
+                  );
                 }
               }}
             >
@@ -145,7 +184,6 @@ function UserListItem({
                 ? "Accept Request"
                 : "Send Request"}{" "}
             </div>
-            q
           </div>
         </div>
       </div>
