+++ packages/message-sdk/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./ChatScreen";
 export * from "./Inbox";
 export * from "./ProfileScreen";
 export * from "./RequestsScreen";
+export * from "./UserList";
